Stat sheet refactor
