+++ apps/web/src/components/dash/admin/events/EventStatsSheet.tsx
@@ -1,38 +1,29 @@
 import React from "react";
 import { getEventStatsOverview } from "@/lib/queries/events";
-
-import { Separator } from "@/components/ui/separator";
+import { StatItemProps } from "@/components/dash/shared/StatItem";
+import StatsSheet from "@/components/dash/shared/StatsSheet";
 
 type Props = {};
 
 async function EventStatsSheet({}: Props) {
 	const stats = await getEventStatsOverview();
-	return (
-		<div className="flex w-fit space-x-4">
-			<div className="flex flex-col p-1">
-				<span className="text-xs text-muted-foreground">
-					Total Events
-				</span>
-				<span className="text-lg font-semibold">
-					{stats.totalEvents}
-				</span>
-			</div>
-			<Separator orientation="vertical" />
-			<div className="flex flex-col p-1">
-				<span className="text-xs text-muted-foreground">This Week</span>
-				<span className="text-lg font-semibold">{stats.thisWeek}</span>
-			</div>
-			<Separator orientation="vertical" />
-			<div className="flex flex-col p-1">
-				<span className="text-xs text-muted-foreground">
-					Past Events
-				</span>
-				<span className="text-lg font-semibold">
-					{stats.pastEvents}
-				</span>
-			</div>
-		</div>
-	);
+
+	const statItems: StatItemProps[] = [
+		{
+			label: "Total Events",
+			value: stats.totalEvents,
+		},
+		{
+			label: "This Week",
+			value: stats.thisWeek,
+		},
+		{
+			label: "Past Events",
+			value: stats.pastEvents,
+		},
+	];
+
+	return <StatsSheet items={statItems} />;
 }
 
 export default EventStatsSheet;
