+++ apps/web/src/components/dash/shared/CheckinStatsSheet.tsx
@@ -1,23 +1,21 @@
 import React from "react";
-import { Separator } from "@/components/ui/separator";
 import { getCheckinStatsOverview } from "@/lib/queries/checkins";
+import { StatItemProps } from "@/components/dash/shared/StatItem";
+import StatsSheet from "@/components/dash/shared/StatsSheet";
 
 type Props = {};
 
 async function CheckinStatsSheet({}: Props) {
 	const stats = await getCheckinStatsOverview();
-	return (
-		<div className="flex w-fit space-x-4">
-			<div className="flex flex-col p-1">
-				<span className="text-xs text-muted-foreground">
-					Total Checkins
-				</span>
-				<span className="text-lg font-semibold">
-					{stats.total_checkins}
-				</span>
-			</div>
-		</div>
-	);
+
+	const statItems: StatItemProps[] = [
+		{
+			label: "Total Checkins",
+			value: stats.total_checkins,
+		},
+	];
+
+	return <StatsSheet items={statItems} />;
 }
 
 export default CheckinStatsSheet;
